Refactor push notification feature list into resuable component
